requirements.txt
Django==2.2.3
PyJWT==1.7.1
onem-py==1.2
requests==2.22.0
